requirements.txt
streamlit>=1.40.0
requests>=2.32.0
Pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
plotly>=5.18.0
